chore: remove unnecessary indirection for regstore
